% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.dispersion.R
\name{ds.dispersion}
\alias{ds.dispersion}
\title{Compute Dispersion Estimate for GLM (Used in Sandwich Estimator)}
\usage{
ds.dispersion(fit, data, datasources = NULL)
}
\arguments{
\item{fit}{A list. Output from `ds.glm`, must include a `formula` and `coefficients`.}

\item{data}{A character string. Name of the data frame stored on the server side.}

\item{datasources}{A list of `DSConnection-class` objects. If `NULL`, attempts to find active connections using `datashield.connections_find()`.}
}
\value{
A numeric value. The estimated dispersion across studies (pooled).
}
\description{
This function computes the dispersion (variance inflation) term for a GLM model
fitted via DataSHIELD. It is used internally for robust standard error estimation,
especially with sandwich (HC-type) estimators.
}
\details{
This function generates predicted values using the GLM link function and then computes
a dispersion estimate by comparing residual and fitted values aggregated across studies.

**Server-side functions called**:
\itemize{
  \item `dispersionDS`
}

 @author Roy Gusinow
}
