% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.eCDF_plot.R
\name{ds.eCDF_plot}
\alias{ds.eCDF_plot}
\title{Plot Empirical Cumulative Distribution Functions (eCDFs) for Matched and Unmatched Data}
\usage{
ds.eCDF_plot(unmatched, matched, formula, len = 40, datasources = NULL)
}
\arguments{
\item{unmatched}{A character string. Name of the unmatched data frame on the server.}

\item{matched}{A character string. Name of the matched data frame on the server.}

\item{formula}{An R formula object specifying the variables. The first variable should be the treatment indicator.}

\item{len}{Integer. Number of interval points used to construct the domain of the eCDF.}

\item{datasources}{A list of `DSConnection-class` objects. If `NULL`, the function attempts to find active connections via `datashield.connections_find()`.}
}
\value{
A combined `ggplot2` object showing eCDF plots.
}
\description{
This function plots eCDFs of variables defined in a formula for both unmatched and matched datasets
using `ds.eCDF` from the DataSHIELD environment. It visually compares the distribution of covariates
between treated and control groups before and after matching.
}
\details{
This function creates server-side subsets for treated and control groups using a binary treatment variable.
It then calls `ds.eCDF` to retrieve the cumulative distribution information and plots the results for each
covariate in the formula using `ggplot2`.
}
\author{
Roy Gusinow
}
