% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.estfun.R
\name{ds.estfun}
\alias{ds.estfun}
\title{This function evaluates the estimating (score) function for each observation in a GLM model.
In the simplest case, this corresponds to the product of the design matrix and residuals, i.e.,
the derivative of the log-likelihood with respect to parameters. It is primarily used for
heteroskedasticity-consistent (HC) variance estimation.}
\usage{
ds.estfun(fit, data, error_type = "HC0", h_diag = "h_diag", datasources = NULL)
}
\arguments{
\item{fit}{A list. Output from `ds.glm`, containing at least `formula`, `coefficients`, and `family`.}

\item{data}{A character string. The name of the data frame on the server-side.}

\item{error_type}{A character string specifying the HC-type estimator to use. One of `"HC0"`, `"HC1"`, `"HC2"`, or `"HC3"`. Default is `"HC0"`.}

\item{h_diag}{A character string. The name of the object on the server that contains the diagonals of the hat matrix. Typically created with `ds.hat_diag()`.}

\item{datasources}{A list of `DSConnection-class` objects. If `NULL`, the function uses `datashield.connections_find()` to detect active connections.}
}
\value{
A list of matrices (one per study) containing scaled estimating functions for each observation and model parameter.
}
\description{
This function evaluates the estimating (score) function for each observation in a GLM model.
In the simplest case, this corresponds to the product of the design matrix and residuals, i.e.,
the derivative of the log-likelihood with respect to parameters. It is primarily used for
heteroskedasticity-consistent (HC) variance estimation.
}
\details{
In DataSHIELD it is not possible to see the data on the servers of the collaborating studies.
It is only possible to get summaries of objects stored on the server-side. This function computes
the estimating function (score function) using predicted values and hat matrix diagonals, for use in
robust variance estimation.

**Server-side functions called**:
\itemize{
  \item `estfunDS`
}

 @author Roy Gusinow
}
