% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.genProp.R
\name{ds.genProp}
\alias{ds.genProp}
\title{Generate Propensity Scores on the Server Side}
\usage{
ds.genProp(
  formula,
  coefficients,
  data,
  newobj = "propscore",
  link = "identity",
  datasources = NULL
)
}
\arguments{
\item{formula}{A formula object. The regression formula used to compute the scores.}

\item{coefficients}{A numeric vector. Coefficients corresponding to the variables in the formula.}

\item{data}{A character string. The name of the data frame on the server side to operate on.}

\item{newobj}{A character string. The name of the new object to be created on the server side. Default is `"propscore"`.}

\item{link}{A character string. The link function used in the regression (e.g., `"logit"`, `"identity"`). Default is `"identity"`.}

\item{datasources}{A list of `DSConnection-class` objects. If `NULL`, the function uses `datashield.connections_find()` to detect active connections.}
}
\value{
No return value. A new object is created on the server side.
}
\description{
Generate Propensity Scores on the Server Side
}
\details{
In DataSHIELD it is not possible to see the data on the servers of the collaborating studies. It is only possible to get summaries of objects stored on the server-side. This function creates a new server-side object containing the predicted values from a linear predictor using a specified regression formula and coefficients.

Server function called: genPropDS
}
\author{
Roy Gusinow
}
