% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.hat_diag.R
\name{ds.hat_diag}
\alias{ds.hat_diag}
\title{Assign Hat Matrix Diagonal to Server-side Object}
\usage{
ds.hat_diag(fit, data, newobj_name = "h_diag", datasources = NULL)
}
\arguments{
\item{fit}{A list. The GLM output object from `ds.glm`, which must contain at least a `formula` and sample size (`Ntotal`).}

\item{data}{A character string. The name of the server-side data frame object.}

\item{newobj_name}{A character string. The name of the new object to be created on the server. Default is `"h_diag"`.}

\item{datasources}{A list of `DSConnection-class` objects. If `NULL`, will attempt to find active connections via `datashield.connections_find()`.}
}
\value{
No return value
}
\description{
Computes and assigns the diagonal of the hat matrix (leverages) from a GLM model to each study server.
These diagonals are useful for influence diagnostics and robust error estimation.
}
\details{
In federated settings using DataSHIELD, hat matrix diagonals cannot be retrieved directly but can be reconstructed
using global covariate means and the global covariance matrix. This function coordinates that process and assigns
the result server-side for later use in robust variance calculations.

**Server-side functions called**:
\itemize{
  \item `hat_diagDS`
}
}
\author{
Roy Gusinow
}
