% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ds.jacobian.R
\name{ds.jacobian}
\alias{ds.jacobian}
\title{This function computes the Jacobian of the average treatment effect (ATE, ATT, or ATC)
with respect to the model parameters in a GLM. It is used in the computation of heteroskedasticity-consistent
variance estimators in federated settings.}
\usage{
ds.jacobian(
  fit,
  data,
  treatment,
  estimand = "ATE",
  comparison = "difference",
  eps = 1e-07,
  datasources = NULL
)
}
\arguments{
\item{fit}{A list. Output from a `ds.glm` model. Must contain required elements like the formula and coefficients.}

\item{data}{A character string. The name of the data frame stored on the server side.}

\item{treatment}{A character string. The name of the treatment variable within `data`.}

\item{estimand}{A character string specifying the estimand to compute. One of `"ATE"`, `"ATT"`, or `"ATC"`. Default is `"ATE"`.}

\item{eps}{A numeric value. The finite difference step size for approximating the Jacobian. Default is `1e-7`.}

\item{datasources}{A list of `DSConnection-class` objects. If `NULL`, the function uses `datashield.connections_find()` to detect active connections.}
}
\value{
A numeric matrix representing the Jacobian of the treatment effect with respect to the model parameters.
}
\description{
This function computes the Jacobian of the average treatment effect (ATE, ATT, or ATC)
with respect to the model parameters in a GLM. It is used in the computation of heteroskedasticity-consistent
variance estimators in federated settings.
}
\details{
In federated analyses, the Jacobian is approximated using finite differences by perturbing
each coefficient in the GLM and recomputing the treatment effect.

Let \eqn{J = \nabla_{\beta} ATE} denote the gradient of the average treatment
effect with respect to the estimated parameters \eqn{\beta_{ATE}}. The
\eqn{k}-th component of this gradient can be obtained numerically using a
finite-difference approximation with step size \eqn{\varepsilon}:

\deqn{
J_k
=
\frac{
ATE(X; \widehat{\beta}_k)
-
ATE(X; \widehat{\beta}_k + \varepsilon)
}{\varepsilon}
}

Because the expression for the average treatment effect is computed in a
federated and secure manner, the corresponding gradient expression is likewise
non-disclosive.

**Server-side functions called**:
\itemize{
  \item None
}
}
\author{
Roy Gusinow
}
