% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{package}
\name{dsMatchingClient-package}
\alias{dsMatchingClient}
\alias{dsMatchingClient-package}
\title{Internal utilities and namespace imports}
\description{
These imports ensure that functions used non-explicitly (e.g., in ggplot2) are available during package checks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/roygusinow/dsMatchingClient}
  \item Report bugs at \url{https://github.com/roygusinow/dsMatchingClient/issues}
}

}
\author{
\strong{Maintainer}: Roy Gusinow \email{rgusinow@uni-bonn.de}

}
\keyword{internal}
